Add source files FrontEnd
